src/main/java/com/library/service/AuthorService.java
package com.library.service;

import com.library.entity.AuthorEntity;
import com.library.repository.AuthorRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class AuthorService {

    @Autowired
    private AuthorRepository authorRepository;

    public List<AuthorEntity> getAllAuthors() {
        return authorRepository.findAll();
    }

    public AuthorEntity getAuthorById(Long id) {
        return authorRepository.findById(id).orElse(null);
    }

    public AuthorEntity createOrUpdateAuthor(AuthorEntity author) {
        return authorRepository.save(author);
    }

    public void deleteAuthor(Long id) {
        authorRepository.deleteById(id);
    }
}
